requirements.txt
opencv-python-headless==4.8.1.78
ultralytics
streamlit
gtts
numpy
Pillow